+++ positivebigint.h
@@ -17,17 +17,17 @@ using namespace std;
 
 class PositiveBigInt{
     private:
-        vector<int> num; 
+        vector<short> num; 
         PositiveBigInt() {}
 
     public:
 
         PositiveBigInt(string s){ 
 
-            int n = s.length(); 
-            int digit; 
+            short n = s.length(); 
+            short digit; 
 
-            for(int i = n-1; i >= 0; i--){
+            for(short i = n-1; i >= 0; i--){
                 if(s[i] >= 48 && s[i] <= 57){
                     digit = s[i] - '0';
                     num.push_back(digit);
@@ -50,13 +50,13 @@ class PositiveBigInt{
 
         PositiveBigInt(const PositiveBigInt& copy){
             
-            int n = copy.num.size();
+            short n = copy.num.size();
 
             while(!(this->num.empty())){
                 this->num.pop_back();
             }
 
-            for(int i=0; i < n; i++){
+            for(short i=0; i < n; i++){
                 this->num.push_back(copy.num[i]);
             }
 
@@ -64,7 +64,7 @@ class PositiveBigInt{
 
 
         string toString();
-        PositiveBigInt betweenMultiplicationAdd(PositiveBigInt& ans, PositiveBigInt& ans2, int i);
+        PositiveBigInt betweenMultiplicationAdd(PositiveBigInt& ans, PositiveBigInt& ans2, short i);
 
         friend ostream& operator<<(ostream& os, const PositiveBigInt &obj);
         void operator=(const PositiveBigInt& obj);
@@ -76,9 +76,9 @@ class PositiveBigInt{
 string PositiveBigInt::toString(){
 
     string ans = "";
-    int n = this->num.size();
+    short n = this->num.size();
 
-    for(int i=n-1; i >= 0; i--){
+    for(short i=n-1; i >= 0; i--){
         ans.push_back(this->num[i] + '0');
     }
 
@@ -89,9 +89,9 @@ string PositiveBigInt::toString(){
 
 ostream& operator<<(ostream& os, const PositiveBigInt &obj){
 
-    int n = obj.num.size();
+    short n = obj.num.size();
     
-    for(int i=n-1; i >= 0; i--){
+    for(short i=n-1; i >= 0; i--){
         os << (obj.num[i]);
     }
 
@@ -101,13 +101,13 @@ ostream& operator<<(ostream& os, const PositiveBigInt &obj){
 
 void PositiveBigInt::operator=(const PositiveBigInt& copy){ 
     
-    int n = copy.num.size();
+    short n = copy.num.size();
 
     while(!(this->num.empty())){
         this->num.pop_back();
     }
 
-    for(int i=0; i < n; i++){
+    for(short i=0; i < n; i++){
         this->num.push_back(copy.num[i]); 
     }
 
@@ -117,8 +117,8 @@ void PositiveBigInt::operator=(const PositiveBigInt& copy){
 PositiveBigInt PositiveBigInt::operator+(PositiveBigInt &obj){
 
     PositiveBigInt ans; 
-    int n = this->num.size(); 
-    int m = obj.num.size(); 
+    short n = this->num.size(); 
+    short m = obj.num.size(); 
     
     if(m > n){
         PositiveBigInt temp(*this);
@@ -129,19 +129,19 @@ PositiveBigInt PositiveBigInt::operator+(PositiveBigInt &obj){
         m = obj.num.size();
     }
 
-    int data;
-    int digit;
-    int carry = 0;
+    short data;
+    short digit;
+    short carry = 0;
     
-    for(int i=0; i < m; i++){
+    for(short i=0; i < m; i++){
         data = this->num[i] + obj.num[i] + carry; 
         digit = data % 10;
         carry = data / 10;
         ans.num.push_back(digit);
     }
 
-    int i = m;
-    int temp = n;
+    short i = m;
+    short temp = n;
     while(i != temp){
         data = this->num[i] + carry;
         digit = data % 10;
@@ -159,16 +159,16 @@ PositiveBigInt PositiveBigInt::operator+(PositiveBigInt &obj){
 } 
 
 
-PositiveBigInt PositiveBigInt::betweenMultiplicationAdd(PositiveBigInt& ans, PositiveBigInt& ans2, int i){
+PositiveBigInt PositiveBigInt::betweenMultiplicationAdd(PositiveBigInt& ans, PositiveBigInt& ans2, short i){
 
-    int n = ans.num.size();
-    int m = ans2.num.size();
+    short n = ans.num.size();
+    short m = ans2.num.size();
 
-    int data;
-    int digit;
-    int carry = 0;
+    short data;
+    short digit;
+    short carry = 0;
 
-    int j=0;
+    short j=0;
     for(; i < n; i++, j++){ 
         data = ans.num[i] + ans2.num[j] + carry; 
         digit = data % 10; 
@@ -196,8 +196,8 @@ PositiveBigInt PositiveBigInt::operator*(PositiveBigInt &obj){
 
 
     PositiveBigInt ans;
-    int n = this->num.size();
-    int m = obj.num.size();
+    short n = this->num.size();
+    short m = obj.num.size();
 
     if( (m == 1 && obj.num[0] == 0) || (n == 1 && this->num[0] == 0)){
         return PositiveBigInt("0");
@@ -213,12 +213,12 @@ PositiveBigInt PositiveBigInt::operator*(PositiveBigInt &obj){
     }
 
 
-    int multiply_data;
-    int multiply_digit;
-    int multiply_carry = 0;
+    short multiply_data;
+    short multiply_digit;
+    short multiply_carry = 0;
     
-    int i = 0;
-    for(int j=0; j < n; j++){
+    short i = 0;
+    for(short j=0; j < n; j++){
 
         multiply_data = (obj.num[i] * this->num[j]) + multiply_carry; 
         multiply_digit = multiply_data % 10;
@@ -232,13 +232,13 @@ PositiveBigInt PositiveBigInt::operator*(PositiveBigInt &obj){
     }
 
 
-    for(int i=1; i < m; i++){
+    for(short i=1; i < m; i++){
         
         multiply_carry = 0;
 
         PositiveBigInt ans2;
 
-        for(int j=0; j < n; j++){
+        for(short j=0; j < n; j++){
 
             multiply_data = (obj.num[i] * this->num[j]) + multiply_carry; 
             multiply_digit = multiply_data % 10;
